# swagger-yml.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: Server real
    url: 0.0.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://app.swaggerhub.com/apis/DianNurdiana-alt/story-six_3/1.0.0
info:
  description: Social Media Apps Project by altera accademy
  version: "1.0.0"
  title: Story six3
tags:
  - name: Authentication
    description: operation comment
  - name: Users
    description: operation user
  - name: Post
    description: operation post
  - name: Comment
    description: operation comment
paths:
  /login:
    post:
      tags:
        - Authentication
      description: fitur untuk login ke sistem
      requestBody:
         content:
            multipart/form-data:
              schema:
                type: object
                properties:                 
                  email:
                    type: string
                  password:
                    type: string
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseLoginSuccess"
              examples:
                responseData1:
                  description: response success 1
                  value:
                    {
                      code: 200,
                      status: "success",
                      message: "Login Success",
                      data:
                        {
                          name: "alex",
                          token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JpemVkIjp0cnVlLCJleHAiOjE2MzgzMjYxODAsInVzZXJJZCI6MSwidXNlcm5hbWUiOiJhZG1pbiJ9.AebFR-oQjUSOMez2ucDWkiMrS2eQIPmcYm5c71qZ_co",
                        },
                    }
        "400":
          description: Failed Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/failedResponse"
              examples:
                responseData1:
                  description: response failed 1
                  value:
                    {
                      code: 400,
                      status: "failed",
                      message: "email or password incorrect",
                    }  
  /register:
    post:
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#"
            example:
              name: alex
              email: alexalttera@mail.com
              password: qwerty123
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  description: response success 1
                  value:
                    {
                      code: 200,
                      status: "success",
                      message: "success operation create data",
                      data:
                        {
                          ID: 39,
                          Nama: "alex",
                          Email: "alexalttera@mail.com",
                          Password: "$2a$10$Kj32oCSvB/ifssRWZinbE..3DurySB2kIeL5FY2BOMLsPSo90n2s2",
                          CreatedAt: "2022-07-26T10:43:46.656+07:00",
                          UpdatedAt: "2022-07-26T10:43:46.656+07:00"
                        }
                    }
        "400":
          description: Failed Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/failedResponse"
              examples:
                responseData1:
                  description: response failed 1
                  value:
                    {
                      code: 400,
                      status: "failed",
                      message: "please make sure all fields are filled in correctly",
                    }
  /users:
    get:
      tags:
        - Users
      description: fitur untuk melihat profil yang login
      security:
        - JWTAuth: []
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  description: response success 1
                  value:
                    {
                      ID: 1,
                      Nama: "alex",
                      Email: "alex@gmail.com",
                      Password: "$2a$10$Kj32oCSvB/ifssRWZinbE..3DurySB2kIeL5FY2BOMLsPSo90n2s2",
                      CreatedAt: 2022-07-26T03:43:46.656Z,
                      UpdatedAt: 2022-07-26T03:43:46.656Z

                    }
        "400":
          description: Failed Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/failedResponse"
              examples:
                responseData1:
                  description: response failed 1
                  value: { code: 400, status: "failed", message: "bad request" }
    put:
      tags:
        - Users
      description: fitur untuk mengupdate profile user yang login
      security:
        - JWTAuth: []
      requestBody:
         content:
            aplication/json:
              schema:
                type: object
                properties:                 
                  New Nama:
                    type: string
                  New Email:
                    type: string
                  New Password:
                    type: string
            multipart/formdata:
              schema:
                type: object
                properties:                 
                  New Nama:
                    type: string
                  New Email:
                    type: string
                  New Password:
                    type: string
      responses:
         "201":
            description: success response
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/successResponse"
                examples:
                  responseData1:
                    description: response success
                    value:
                      {
                        code: 201,
                        status: "success",
                        message: "Successfully updated data",
                      }
         "400":
            description: Failed Response
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/failedResponse"
                examples:
                  responseData1:
                    description: response failed 1
                    value:
                      {
                        code: 400,
                        status: "failed",
                        message: "Please make sure all fields are filled in correctly",
                      }  
    delete:
      tags:
        - Users
      security:
        - JWTAuth: []
      parameters:
        - name: iduser
          in: query
          schema:
            type: integer
            example: 1
          required: true
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  value:
                    code: 200
                    status: "success"
                    message: "Successful Operation"
        "400":
          description: Failed Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/failedResponse"
              examples:
                responseData1:
                  description: response failed 1
                  value: { code: 400, status: "failed", message: "Bad Request" }
   
  /posts:
    get:
      tags:
        - Post
      description: fitur untuk melihat profil yang login
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  description: response success
                  value:
                    {
                      ID: 1,
                      Content: "ini adalah status pertamaku di story six3",
                      Image: "https://gudanggroup3.s3.amazonaws.com/2022-07-06%2006%3A40%3A47.jpg",
                      CreatedAt: 0001-01-01T00:00:00Z,
                      UpdatedAt: 0001-01-01T00:00:00Z,
                        user: 
                        {
                          ID: 23,
                          Nama: "alex"
                        }
                    }
        "400":
          description: Failed Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/failedResponse"
              examples:
                responseData1:
                  description: response failed
                  value: { code: 400, status: "failed", message: "bad request" }
    post:
      tags:
        - Post
      description: fitur untuk membuat postingan
      security:
        - JWTAuth: []
      requestBody:
        content:
          multipart/form-data:
              schema:
                type: object
                properties: 
                  Content:
                    type: string              
                  Image:
                    type: string
                    format: binary
      responses:
        '200':
            description: Success Response
            content:
              application/json:
                schema:
                  $ref: "#"
                examples:
                  response data:
                    description: response success 1
                    value:
                      {
                      ID: 1,
                      Content: "ini adalah status pertamaku di story six3",
                      Image: "https://gudanggroup3.s3.amazonaws.com/2022-07-06%2006%3A40%3A47.jpg",
                      CreatedAt: 0001-01-01T00:00:00Z,
                      UpdatedAt: 0001-01-01T00:00:00Z,
                        user: 
                        {
                          ID: 23,
                          Nama: "alex"
                        }
                      }
        '500':
          description: "Internal server error"
    put:
      tags:
        - Post
      description: fitur untuk mengupdate postingan user yang login
      security:
        - JWTAuth: []
      requestBody:
         content:
            multipart/form-data:
              schema:
                type: object
                properties: 
                  Content:
                    type: string              
                  Image:
                    type: string
                    format: binary
      responses:
        "200":
            description: Success Response
            content:
              application/json:
                schema:
                  $ref: "#"
                examples:
                  response data:
                    description: response success 1
                    value:
                      {
                      ID: 1,
                      Content: "ini adalah status pertamaku di story six3",
                      Image: "https://gudanggroup3.s3.amazonaws.com/2022-07-06%2006%3A40%3A47.jpg",
                      CreatedAt: 0001-01-01T00:00:00Z,
                      UpdatedAt: 0001-01-01T00:00:00Z,
                        user: 
                        {
                          ID: 23,
                          Nama: "alex"
                        }
                      }
        "400":
            description: Failed Response
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/failedResponse"
                examples:
                  responseData1:
                    description: response failed 1
                    value:
                      {
                        code: 400,
                        status: "failed",
                        message: "Please make sure all fields are filled in correctly",
                      }
        '500':
          description: "Internal server error"  
    delete:
      tags:
        - Post
      description: fitur menghapus postingan
      security:
        - JWTAuth: []
      parameters:
        - name: idpost
          in: query
          schema:
            type: integer
            example: 1
          required: true
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  value:
                    code: 200
                    status: "success"
                    message: "Successful Operation"
        "400":
          description: Failed Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/failedResponse"
              examples:
                responseData1:
                  description: response failed 1
                  value: { code: 400, status: "failed", message: "Bad Request" }  
 
  /comments:
    get:
      tags:
        - Comment
      description: fitur menampilkan semua komentar
      security:
        - JWTAuth: []
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  description: response success
                  value:
                   {
                      ID: 1,
                      Comment: ini adalah komen pertamaku,
                      CreatedAt: 0001-01-01T00:00:00Z,
                      id_user: 2,
                      Posting:
                        {
                        ID: 1,
                        Content: "ini adalah status pertamaku di story six3",
                        Image: "https://gudanggroup3.s3.amazonaws.com/2022-07-06%2006%3A40%3A47.jpg",
                        }
                    }
        '500':
          description: "Internal server error"
    post:
      tags:
        - Comment
      description: fitur untuk menambahkan komentar
      security:
        - JWTAuth: []
      requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  comment:
                    type: string     
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  description: response success
                  value:
                    {
                      ID: 1,
                      Comment: ini adalah komen pertamaku,
                      CreatedAt: 0001-01-01T00:00:00Z,
                      id_user: 2,
                      Posting:
                        {
                        ID: 1,
                        Content: "ini adalah status pertamaku di story six3",
                        Image: "https://gudanggroup3.s3.amazonaws.com/2022-07-06%2006%3A40%3A47.jpg",
                        }
                    }
        '500':
          description: "Internal server error"
    delete:
      tags:
        - Comment
      description: fitur untuk menghapus komentar
      security:
        - JWTAuth: []

      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: "#"
              examples:
                response data:
                  description: response success
                  value: { code: 200, status: "success", message: "Successfully operation" } 
        "400":
          description: Failed Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/failedResponse"
              examples:
                responseData1:
                  description: response failed 1
                  value: { code: 400, status: "failed", message: "Bad Request" } 

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    modelUser:
      type: object
      properties:
        "ID":
          type: integer
          format: int32
        "Nama":
          type: string
        "Email":
          type: string
        "Password": 
          type: string
          format: password
    requestLogin:
      type: object
      properties:
        "email":
          type: string
          minLength: 1
        "password":
          type: string
          minLength: 1
    responseLoginSuccess:
      type: object
      properties:
        {
          "code": { type: number },
          "message": { type: string },
          "status": { type: string },
          "data": { type: object },
        }
    successResponse:
      type: object
      properties:
        {
          "code": { type: number },
          "message": { type: string },
          "status": { type: string },
        }
    failedResponse:
      type: object
      properties:
        {
          "code": { type: number },
          "message": { type: string },
          "status": { type: string },
        }